Rename panelView2.tsx
