+++ packages/core/src/components/panel-stack2/panelView.tsx
